Edit the post.
I've tweaked the wording quite a bit, to make it more succinct, and I
reorganized the sections somewhat, so that we first tell a story and then
reveal our final configuration. I also decided to drop the explanation of each
`connect_args` param since people can easily look that up in the docs.